+++ controller/api/public/http_server.go
@@ -7,33 +7,21 @@ import (
 
 	"github.com/golang/protobuf/proto"
 	destinationPb "github.com/linkerd/linkerd2-proxy-api/go/destination"
-	healthcheckPb "github.com/linkerd/linkerd2/controller/gen/common/healthcheck"
-	publicPb "github.com/linkerd/linkerd2/controller/gen/public"
+	pb "github.com/linkerd/linkerd2/controller/gen/public"
 	"github.com/linkerd/linkerd2/controller/k8s"
 	"github.com/linkerd/linkerd2/pkg/prometheus"
 	"github.com/linkerd/linkerd2/pkg/protohttp"
-	pb "github.com/linkerd/linkerd2/viz/metrics-api/gen/viz"
-	promApi "github.com/prometheus/client_golang/api"
-	promv1 "github.com/prometheus/client_golang/api/prometheus/v1"
 	log "github.com/sirupsen/logrus"
 	"google.golang.org/grpc/metadata"
 )
 
 var (
-	gatewaysPath     = fullURLPathFor("Gateways")
-	statSummaryPath  = fullURLPathFor("StatSummary")
-	topRoutesPath    = fullURLPathFor("TopRoutes")
-	versionPath      = fullURLPathFor("Version")
-	listPodsPath     = fullURLPathFor("ListPods")
-	listServicesPath = fullURLPathFor("ListServices")
-	selfCheckPath    = fullURLPathFor("SelfCheck")
-	edgesPath        = fullURLPathFor("Edges")
-	destGetPath      = fullURLPathFor("DestinationGet")
+	versionPath = fullURLPathFor("Version")
+	destGetPath = fullURLPathFor("DestinationGet")
 )
 
 type handler struct {
-	publicGRPCServer PublicAPIServer
-	grpcServer       VizAPIServer
+	grpcServer Server
 }
 
 func (h *handler) ServeHTTP(w http.ResponseWriter, req *http.Request) {
@@ -48,22 +36,8 @@ func (h *handler) ServeHTTP(w http.ResponseWriter, req *http.Request) {
 
 	// Serve request
 	switch req.URL.Path {
-	case gatewaysPath:
-		h.handleGateways(w, req)
-	case statSummaryPath:
-		h.handleStatSummary(w, req)
-	case topRoutesPath:
-		h.handleTopRoutes(w, req)
 	case versionPath:
 		h.handleVersion(w, req)
-	case listPodsPath:
-		h.handleListPods(w, req)
-	case listServicesPath:
-		h.handleListServices(w, req)
-	case selfCheckPath:
-		h.handleSelfCheck(w, req)
-	case edgesPath:
-		h.handleEdges(w, req)
 	case destGetPath:
 		h.handleDestGet(w, req)
 	default:
@@ -72,163 +46,15 @@ func (h *handler) ServeHTTP(w http.ResponseWriter, req *http.Request) {
 
 }
 
-func (h *handler) handleGateways(w http.ResponseWriter, req *http.Request) {
-	var protoRequest pb.GatewaysRequest
-
-	err := protohttp.HTTPRequestToProto(req, &protoRequest)
-	if err != nil {
-		protohttp.WriteErrorToHTTPResponse(w, err)
-		return
-	}
-
-	rsp, err := h.grpcServer.Gateways(req.Context(), &protoRequest)
-	if err != nil {
-		protohttp.WriteErrorToHTTPResponse(w, err)
-		return
-	}
-	err = protohttp.WriteProtoToHTTPResponse(w, rsp)
-	if err != nil {
-		protohttp.WriteErrorToHTTPResponse(w, err)
-		return
-	}
-}
-
-func (h *handler) handleStatSummary(w http.ResponseWriter, req *http.Request) {
-	var protoRequest pb.StatSummaryRequest
-
-	err := protohttp.HTTPRequestToProto(req, &protoRequest)
-	if err != nil {
-		protohttp.WriteErrorToHTTPResponse(w, err)
-		return
-	}
-
-	rsp, err := h.grpcServer.StatSummary(req.Context(), &protoRequest)
-	if err != nil {
-		protohttp.WriteErrorToHTTPResponse(w, err)
-		return
-	}
-	err = protohttp.WriteProtoToHTTPResponse(w, rsp)
-	if err != nil {
-		protohttp.WriteErrorToHTTPResponse(w, err)
-		return
-	}
-}
-
-func (h *handler) handleEdges(w http.ResponseWriter, req *http.Request) {
-	var protoRequest pb.EdgesRequest
-
-	err := protohttp.HTTPRequestToProto(req, &protoRequest)
-	if err != nil {
-		protohttp.WriteErrorToHTTPResponse(w, err)
-		return
-	}
-
-	rsp, err := h.grpcServer.Edges(req.Context(), &protoRequest)
-	if err != nil {
-		protohttp.WriteErrorToHTTPResponse(w, err)
-		return
-	}
-	err = protohttp.WriteProtoToHTTPResponse(w, rsp)
-	if err != nil {
-		protohttp.WriteErrorToHTTPResponse(w, err)
-		return
-	}
-}
-
-func (h *handler) handleTopRoutes(w http.ResponseWriter, req *http.Request) {
-	var protoRequest pb.TopRoutesRequest
-
-	err := protohttp.HTTPRequestToProto(req, &protoRequest)
-	if err != nil {
-		protohttp.WriteErrorToHTTPResponse(w, err)
-		return
-	}
-
-	rsp, err := h.grpcServer.TopRoutes(req.Context(), &protoRequest)
-	if err != nil {
-		protohttp.WriteErrorToHTTPResponse(w, err)
-		return
-	}
-	err = protohttp.WriteProtoToHTTPResponse(w, rsp)
-	if err != nil {
-		protohttp.WriteErrorToHTTPResponse(w, err)
-		return
-	}
-}
-
 func (h *handler) handleVersion(w http.ResponseWriter, req *http.Request) {
-	var protoRequest publicPb.Empty
+	var protoRequest pb.Empty
 	err := protohttp.HTTPRequestToProto(req, &protoRequest)
 	if err != nil {
 		protohttp.WriteErrorToHTTPResponse(w, err)
 		return
 	}
 
-	rsp, err := h.publicGRPCServer.Version(req.Context(), &protoRequest)
-	if err != nil {
-		protohttp.WriteErrorToHTTPResponse(w, err)
-		return
-	}
-
-	err = protohttp.WriteProtoToHTTPResponse(w, rsp)
-	if err != nil {
-		protohttp.WriteErrorToHTTPResponse(w, err)
-		return
-	}
-}
-
-func (h *handler) handleSelfCheck(w http.ResponseWriter, req *http.Request) {
-	var protoRequest healthcheckPb.SelfCheckRequest
-	err := protohttp.HTTPRequestToProto(req, &protoRequest)
-	if err != nil {
-		protohttp.WriteErrorToHTTPResponse(w, err)
-		return
-	}
-
-	rsp, err := h.grpcServer.SelfCheck(req.Context(), &protoRequest)
-	if err != nil {
-		protohttp.WriteErrorToHTTPResponse(w, err)
-		return
-	}
-
-	err = protohttp.WriteProtoToHTTPResponse(w, rsp)
-	if err != nil {
-		protohttp.WriteErrorToHTTPResponse(w, err)
-		return
-	}
-}
-
-func (h *handler) handleListPods(w http.ResponseWriter, req *http.Request) {
-	var protoRequest pb.ListPodsRequest
-	err := protohttp.HTTPRequestToProto(req, &protoRequest)
-	if err != nil {
-		protohttp.WriteErrorToHTTPResponse(w, err)
-		return
-	}
-
-	rsp, err := h.grpcServer.ListPods(req.Context(), &protoRequest)
-	if err != nil {
-		protohttp.WriteErrorToHTTPResponse(w, err)
-		return
-	}
-
-	err = protohttp.WriteProtoToHTTPResponse(w, rsp)
-	if err != nil {
-		protohttp.WriteErrorToHTTPResponse(w, err)
-		return
-	}
-}
-
-func (h *handler) handleListServices(w http.ResponseWriter, req *http.Request) {
-	var protoRequest pb.ListServicesRequest
-
-	err := protohttp.HTTPRequestToProto(req, &protoRequest)
-	if err != nil {
-		protohttp.WriteErrorToHTTPResponse(w, err)
-		return
-	}
-
-	rsp, err := h.grpcServer.ListServices(req.Context(), &protoRequest)
+	rsp, err := h.grpcServer.Version(req.Context(), &protoRequest)
 	if err != nil {
 		protohttp.WriteErrorToHTTPResponse(w, err)
 		return
@@ -256,7 +82,7 @@ func (h *handler) handleDestGet(w http.ResponseWriter, req *http.Request) {
 	}
 
 	server := destinationServer{streamServer{w: flushableWriter, req: req}}
-	err = h.publicGRPCServer.Get(&protoRequest, server)
+	err = h.grpcServer.Get(&protoRequest, server)
 	if err != nil {
 		protohttp.WriteErrorToHTTPResponse(w, err)
 		return
@@ -302,30 +128,19 @@ func fullURLPathFor(method string) string {
 // NewServer creates a Public API HTTP server.
 func NewServer(
 	addr string,
-	prometheusClient promApi.Client,
 	destinationClient destinationPb.DestinationClient,
 	k8sAPI *k8s.API,
 	controllerNamespace string,
 	clusterDomain string,
-	ignoredNamespaces []string,
 ) *http.Server {
 
-	var promAPI promv1.API
-	if prometheusClient != nil {
-		promAPI = promv1.NewAPI(prometheusClient)
-	}
-
-	grpcServer := newGrpcServer(
-		promAPI,
-		destinationClient,
-		k8sAPI,
-		controllerNamespace,
-		clusterDomain,
-		ignoredNamespaces,
-	)
 	baseHandler := &handler{
-		publicGRPCServer: grpcServer,
-		grpcServer:       grpcServer,
+		grpcServer: newGrpcServer(
+			destinationClient,
+			k8sAPI,
+			controllerNamespace,
+			clusterDomain,
+		),
 	}
 
 	instrumentedHandler := prometheus.WithTelemetry(baseHandler)
