+++ pkg/public/api.go
@@ -7,14 +7,13 @@ import (
 	"time"
 
 	"github.com/linkerd/linkerd2/controller/api/public"
-	publicPb "github.com/linkerd/linkerd2/controller/gen/public"
+	pb "github.com/linkerd/linkerd2/controller/gen/public"
 	"github.com/linkerd/linkerd2/pkg/healthcheck"
 	"github.com/linkerd/linkerd2/pkg/k8s"
-	pb "github.com/linkerd/linkerd2/viz/metrics-api/gen/viz"
 )
 
 // RawPublicAPIClient creates a raw public API client with no validation.
-func RawPublicAPIClient(ctx context.Context, kubeAPI *k8s.KubernetesAPI, controlPlaneNamespace string, apiAddr string) (publicPb.ApiClient, error) {
+func RawPublicAPIClient(ctx context.Context, kubeAPI *k8s.KubernetesAPI, controlPlaneNamespace string, apiAddr string) (pb.ApiClient, error) {
 	if apiAddr != "" {
 		return public.NewInternalPublicClient(controlPlaneNamespace, apiAddr)
 	}
@@ -22,36 +21,19 @@ func RawPublicAPIClient(ctx context.Context, kubeAPI *k8s.KubernetesAPI, control
 	return public.NewExternalPublicClient(ctx, controlPlaneNamespace, kubeAPI)
 }
 
-// RawVizAPIClient creates a raw viz API client with no validation.
-func RawVizAPIClient(ctx context.Context, kubeAPI *k8s.KubernetesAPI, controlPlaneNamespace string, apiAddr string) (pb.ApiClient, error) {
-	if apiAddr != "" {
-		return public.NewInternalClient(controlPlaneNamespace, apiAddr)
-	}
-
-	return public.NewExternalClient(ctx, controlPlaneNamespace, kubeAPI)
-}
-
 // CheckPublicAPIClientOrExit builds a new Public API client and executes default status
 // checks to determine if the client can successfully perform cli commands. If the
 // checks fail, then CLI will print an error and exit.
-func CheckPublicAPIClientOrExit(hcOptions healthcheck.Options) public.PublicAPIClient {
+func CheckPublicAPIClientOrExit(hcOptions healthcheck.Options) public.Client {
 	hcOptions.RetryDeadline = time.Time{}
 	return CheckPublicAPIClientOrRetryOrExit(hcOptions, false)
-}
-
-// CheckVizAPIClientOrExit builds a new Viz API client and executes default status
-// checks to determine if the client can successfully perform cli commands. If the
-// checks fail, then CLI will print an error and exit.
-func CheckVizAPIClientOrExit(hcOptions healthcheck.Options) public.VizAPIClient {
-	hcOptions.RetryDeadline = time.Time{}
-	return CheckVizAPIClientOrRetryOrExit(hcOptions, false)
 }
 
 // CheckPublicAPIClientOrRetryOrExit builds a new Public API client and executes status
 // checks to determine if the client can successfully connect to the API. If the
 // checks fail, then CLI will print an error and exit. If the hcOptions.retryDeadline
 // param is specified, then the CLI will print a message to stderr and retry.
-func CheckPublicAPIClientOrRetryOrExit(hcOptions healthcheck.Options, apiChecks bool) public.PublicAPIClient {
+func CheckPublicAPIClientOrRetryOrExit(hcOptions healthcheck.Options, apiChecks bool) public.Client {
 	checks := []healthcheck.CategoryID{
 		healthcheck.KubernetesAPIChecks,
 		healthcheck.LinkerdControlPlaneExistenceChecks,
@@ -65,26 +47,6 @@ func CheckPublicAPIClientOrRetryOrExit(hcOptions healthcheck.Options, apiChecks 
 
 	hc.RunChecks(exitOnError)
 	return hc.PublicAPIClient()
-}
-
-// CheckVizAPIClientOrRetryOrExit builds a new Viz API client and executes status
-// checks to determine if the client can successfully connect to the API. If the
-// checks fail, then CLI will print an error and exit. If the hcOptions.retryDeadline
-// param is specified, then the CLI will print a message to stderr and retry.
-func CheckVizAPIClientOrRetryOrExit(hcOptions healthcheck.Options, apiChecks bool) public.VizAPIClient {
-	checks := []healthcheck.CategoryID{
-		healthcheck.KubernetesAPIChecks,
-		healthcheck.LinkerdControlPlaneExistenceChecks,
-	}
-
-	if apiChecks {
-		checks = append(checks, healthcheck.LinkerdAPIChecks)
-	}
-
-	hc := healthcheck.NewHealthChecker(checks, &hcOptions)
-
-	hc.RunChecks(exitOnError)
-	return hc.VizAPIClient()
 }
 
 func exitOnError(result *healthcheck.CheckResult) {
