+++ viz/metrics-api/stat_summary.go
@@ -1,4 +1,4 @@
-package public
+package api
 
 import (
 	"context"
