+++ viz/metrics-api/top_routes.go
@@ -1,4 +1,4 @@
-package public
+package api
 
 import (
 	"context"
