+++ viz/metrics-api/top_routes_test.go
@@ -1,4 +1,4 @@
-package public
+package api
 
 import (
 	"context"
